refactor: extract MentionService for @user/#channel mention clicks
The identical onMentionClick/showProfileOrChannel/caseUser/caseChannel
blocks in message-template and chat-thread now live once in
shared/services/mention; both components delegate.
