+++ src/app/features/app_chat/components/message/message-template.component.ts
@@ -1,6 +1,5 @@
 import { CommonModule } from '@angular/common';
 import { Component, computed, inject, input, Input } from '@angular/core';
-import { UserService } from '../../../../shared/services/user/shared.service';
 import { FireServiceService } from '../../../../shared/services/firebase/fire-service.service';
 import { MatIconModule } from '@angular/material/icon';
 
@@ -12,7 +11,7 @@ import emojiData from 'unicode-emoji-json';
 
 import { LinkifyPipe } from '../../../pipes/linkify.pipe';
 import { UserStore } from '../../../../shared/services/user/user-store';
-import { ChannelService } from '../../../app_channel/services/channel/channel.service';
+import { MentionService } from '../../../../shared/services/mention/mention.service';
 
 import { DialogReciverComponent } from '../../../dialogs/dialog-reciver/dialog-reciver.component';
 import { ChannelMessage } from '../../models/channel-message/channel-message';
@@ -32,7 +31,7 @@ export class MessageTemplateComponent {
   private dialog = inject(MatDialog);
   public navigationService: NavigationService = inject(NavigationService);
   private userStore: UserStore = inject(UserStore);
-  private channelService: ChannelService = inject(ChannelService);
+  private mentionService: MentionService = inject(MentionService);
   menuOpen: boolean = false;
   reactionMenuOpen: boolean = false;
   reactionMenuOpenInFooter: boolean = false;
@@ -266,66 +265,9 @@ export class MessageTemplateComponent {
   }
 
   /**
-   * Handles click events on mention buttons within a message.
-   * Determines whether a mention button was clicked and, if so,
-   * extracts its symbol (@ or #) and the associated name,
-   * then delegates to navigation logic.
-   *
-   * @param event - The MouseEvent triggered by the click.
+   * Delegates clicks inside the rendered message to the MentionService.
    */
   onMentionClick(event: MouseEvent | TouchEvent) {
-    const btn = (event.target as HTMLElement).closest('.tag-btn');
-    if (!btn) return;
-
-    const fullTag = btn.textContent?.trim();
-    if (!fullTag) return;
-
-    const symbol = fullTag.charAt(0);
-    const name = fullTag.slice(1).trim();
-    this.showProfileOrChannel(symbol, name);
-  }
-
-  /**
-   * Routes the click on a mention based on its symbol.
-   * If the symbol is '@', performs a user lookup;
-   * if '#', performs a channel lookup.
-   *
-   * @param symbol - The mention symbol, either '@' or '#'.
-   * @param name - The username or channel name to lookup.
-   */
-  async showProfileOrChannel(symbol: string, name: string) {
-    switch (symbol) {
-      case '@':
-        await this.caseUser(name);
-        break;
-
-      case '#':
-        await this.caseChannel(name);
-        break;
-    }
-  }
-
-  /**
-   * Looks up a user in Firestore by their full name,
-   * sets the navigation service’s receiver ID to the found user’s document ID,
-   * and switches the UI to a direct chat view.
-   *
-   * @param name - The user’s full name to query.
-   */
-  async caseUser(name: string) {
-    const user = await this.userStore.findUserByDisplayName(name);
-    if (user) this.navigationService.selectDirectMessageRecipient(user.id);
-  }
-
-  /**
-   * Looks up a channel in Firestore by its name,
-   * sets the navigation service’s receiver ID to the found channel’s document ID,
-   * and switches the UI to the channel view.
-   *
-   * @param name - The channel’s name to query.
-   */
-  async caseChannel(name: string) {
-    const channel = await this.channelService.findChannelByName(name);
-    if (channel?.id) this.navigationService.selectChannel(channel.id);
+    this.mentionService.handleMentionClick(event);
   }
 }
